+++ toucan_connectors/google_credentials.py
@@ -13,16 +13,16 @@ class GoogleCredentials(BaseModel):
         'service_account', title='Service account', description=CREDENTIALS_INFO_MESSAGE
     )
     # On service account MODE :
-    project_id: str = Field(None, title='Project ID', description=CREDENTIALS_INFO_MESSAGE)
-    private_key_id: str = Field(None, title='Private Key ID', description=CREDENTIALS_INFO_MESSAGE)
+    project_id: str = Field(..., title='Project ID', description=CREDENTIALS_INFO_MESSAGE)
+    private_key_id: str = Field(..., title='Private Key ID', description=CREDENTIALS_INFO_MESSAGE)
     private_key: str = Field(
-        None,
+        ...,
         title='Private Key',
         description=f'A private key in the form '
         f'"-----BEGIN PRIVATE KEY-----\\nXXX...XXX\\n-----END PRIVATE KEY-----\\n". {CREDENTIALS_INFO_MESSAGE}',
     )
-    client_email: str = Field(None, title='Client email', description=CREDENTIALS_INFO_MESSAGE)
-    client_id: str = Field(None, title='Client ID', description=CREDENTIALS_INFO_MESSAGE)
+    client_email: str = Field(..., title='Client email', description=CREDENTIALS_INFO_MESSAGE)
+    client_id: str = Field(..., title='Client ID', description=CREDENTIALS_INFO_MESSAGE)
     auth_uri: HttpUrl = Field(
         'https://accounts.google.com/o/oauth2/auth',
         title='Authentication URI',
@@ -39,7 +39,7 @@ class GoogleCredentials(BaseModel):
         description=f'{CREDENTIALS_INFO_MESSAGE}. You should not need to change the default value.',
     )
     client_x509_cert_url: HttpUrl = Field(
-        None,
+        ...,
         title='Client X509 certification URL',
         description=CREDENTIALS_INFO_MESSAGE,
     )